+++ lib/lokalise_rails/railtie.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 module LokaliseRails
+  # Load Rake tasks
   class Railtie < Rails::Railtie
     rake_tasks do
       load 'tasks/lokalise_rails_tasks.rake'