+++ spec/dummy/config/lokalise_rails.rb
@@ -1,5 +1,6 @@
-LokaliseRails::GlobalConfig.config do |c|
-  c.api_token = ENV['LOKALISE_API_TOKEN']
-  c.project_id = ENV['LOKALISE_PROJECT_ID']
+# frozen_string_literal: true
 
+LokaliseRails::GlobalConfig.config do |c|
+  c.api_token = ENV.fetch('LOKALISE_API_TOKEN', nil)
+  c.project_id = ENV.fetch('LOKALISE_PROJECT_ID', nil)
 end